TUC-159 Adding a treatment-model compatibility checker
This checker checks if the formulation and routes of the treatment
are available in the drug model

	modified:   ../../make/qtcreator/tucucore.pri
	modified:   ../../src/tucucli/clicomputer.cpp
	modified:   ../../src/tucucore/computingcomponent.cpp
	modified:   ../../src/tucucore/computingresponseexport.cpp
	modified:   ../../src/tucucore/computingservice/computingresult.h
	modified:   ../../src/tucucore/dosage.cpp
	modified:   ../../src/tucucore/dosage.h
	modified:   ../../src/tucucore/drugmodel/formulationandroute.cpp
	modified:   ../../src/tucucore/drugmodel/formulationandroute.h
	modified:   ../../src/tucucore/drugmodelrepository.cpp
	modified:   ../../src/tucucore/makefile
	new file:   ../../src/tucucore/treatmentdrugmodelcompatibilitychecker.cpp
	new file:   ../../src/tucucore/treatmentdrugmodelcompatibilitychecker.h
	modified:   ../../src/tucuquery/querytocoreextractor.cpp
	modified:   ../../src/tucuquery/querytocoreextractor.h
	modified:   ../tucucore/drugmodels/test_drug_vancomycin.h

+++ src/tucucore/computingresponseexport.cpp
@@ -1,11 +1,9 @@
 #include "computingresponseexport.h"
 
 #include <iostream>
+#include <sstream>
 #include <fstream>
-#include<iostream>
-#include<sstream>
-#include<fstream>
-#include<iomanip>
+#include <iomanip>
 
 namespace Tucuxi {
 namespace Core {
@@ -57,7 +55,7 @@ bool ComputingResponseExport::exportToFiles(const ComputingResponse &_computingR
                 for (size_t i = 0; i < prediction->getPercentileData(0)[cycleIndex].m_concentrations[0].size(); i++) {
                     file <<  (prediction->getPercentileData(0)[cycleIndex].m_start.toSeconds() - firstTime) / 3600.0
                              + prediction->getPercentileData(0)[cycleIndex].m_times[0][i] << "\t";
-                    for (size_t percIndex = 0; percIndex < prediction->getNbRanks(); percIndex ++) {
+                    for (unsigned int percIndex = 0; percIndex < prediction->getNbRanks(); percIndex ++) {
                         file  << prediction->getPercentileData(percIndex)[cycleIndex].m_concentrations[0][i] << "\t";
                     }
                     file << std::endl;
